little fixes

+++ src/servus/core/external/filer/general.go
@@ -106,10 +106,12 @@ func ProcessFromForm(request *http.Request, formKey string, tempDir string) (dat
 
 	defer func() {
 		// always close.
-		_ = tempFile.Close()
-		if err != nil {
-			// delete if something goes wrong.
-			_ = os.Remove(tempFile.Name())
+		if tempFile != nil {
+			_ = tempFile.Close()
+			if err != nil {
+				// delete if something goes wrong.
+				_ = os.Remove(tempFile.Name())
+			}
 		}
 	}()
